GameState
Added the 'GameState' class.
This controls which state the game is currently in, like the menu or playing the game.

Currently when the window loads it goes to the menu and waits for the play button to be pressed. Once the button is pressed it switches to the playing state where the game runs normaly. While playing it is checking if all the bricks are gone. If they are it will switch to the all bricks destroyed state in prep for something after.